feat: Update homepage to create a new chat and redirect to chat page
Refactor the homepage component to include a new functionality for creating a new chat. This functionality uses the useMutation hook from the Convex library to store the user data and obtain a chat ID. After successfully creating the chat, the user is redirected to the chat page using the useRouter hook from Next.js.

This commit also includes some minor styling changes to the homepage component.