feat: Add UpgradePlanButton component to Sidebar
Add the UpgradePlanButton component to the Sidebar component in order to provide users with the option to upgrade their plan. This button will be displayed alongside the NewChatButton and ChatList components. The UpgradePlanButton component will be responsible for handling the logic and UI for upgrading the user's plan.

This commit adds the UpgradePlanButton component to the Sidebar component.